# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  legocy-go_internal_delievery_http_resources_lego.LegoSetResponse:
    properties:
      id:
        type: integer
      n_pieces:
        type: integer
      name:
        type: string
      number:
        type: integer
      series:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSeriesResponse'
    type: object
  legocy-go_internal_delievery_http_resources_marketplace.MarketItemRequest:
    properties:
      description:
        type: string
      lego_set_id:
        type: integer
      location:
        type: string
      price:
        type: number
      set_state:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_marketplace.MarketItemResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      lego_set:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_lego.LegoSetResponse'
      location:
        type: string
      price:
        type: number
      seller:
        $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.UserDetailResponse'
      set_state:
        type: string
      status:
        type: string
    type: object
...
  legocy-go_internal_delievery_http_resources_users.JWTRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_users.JWTResponse:
    properties:
      access:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_users.UserDetailResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: integer
      username:
        type: string
    type: object
...
  legocy-go_internal_delievery_http_resources_users_admin.AdminRegistrationRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_users_admin.AdminRegistrationResponse:
    properties:
      email:
        type: string
      role:
        type: integer
      username:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_users_admin.UserAdminDetailResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: integer
      username:
        type: string
    type: object
  legocy-go_internal_delievery_http_resources_users_admin.UserAdminUpdateRequest:
    properties:
      email:
        type: string
      role:
        description: Admin/User/etc.
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: LEGOcy is a marketplace for LEGO lovers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LEGOcy API
  version: "1.0"
paths:
...
  /admin/users/:
    get:
      operationId: list_users_admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.UserAdminDetailResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get Users (Admin)
      tags:
      - users_admin
  /admin/users/{userId}:
    delete:
      operationId: delete_user_admin
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - users_admin
    get:
      operationId: detail_user_admin
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.UserAdminDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get User by ID (Admin)
      tags:
      - users_admin
    put:
      operationId: update_user
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.UserAdminUpdateRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.UserAdminDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - users_admin
  /admin/users/register:
    post:
      operationId: create-admin
      parameters:
      - description: reg request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.AdminRegistrationRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users_admin.AdminRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create Admin
      tags:
      - users_admin
...
  /market-items/:
    get:
      operationId: list_market_items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get Market Items
      tags:
      - market_items
    post:
      operationId: create_market_item
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace.MarketItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Create Market Item
      tags:
      - market_items
  /market-items/{itemID}:
    get:
      operationId: detail_market_item
      parameters:
      - description: item ID
        in: path
        name: itemID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace.MarketItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get Market Item
      tags:
      - market_items
    put:
      operationId: update_market_item
      parameters:
      - description: item ID
        in: path
        name: itemID
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace.MarketItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_marketplace.MarketItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Update Market Item
      tags:
      - market_items
  /market-items/{itemId}:
    delete:
      operationId: delete_market_item
      parameters:
      - description: item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Delete Market Item
      tags:
      - market_items
  /market-items/authorized/:
    get:
      operationId: list_market_items_authorized
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: Get Market Items Authorized
      tags:
      - market_items
...
  /users/auth/token:
    post:
      operationId: create-jwt
      parameters:
      - description: jwt request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.JWTRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.JWTResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: generate jwt token
      tags:
      - authentication
...
  /users/images/download:
    get:
      operationId: download_user_image
      parameters:
      - description: filepath
        in: query
        name: fp
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Download User Image
      tags:
      - users_images
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"